Chore: Split tests into several files according to the functionality they're testing.
